Upgrade to gatsby-theme-ocular
